+++ drake/systems/plants/RigidBody.h
@@ -78,6 +78,24 @@ class DRAKERBM_EXPORT RigidBody {
    */
   const DrakeJoint& getJoint() const;
 
+  /**
+   * Sets the parent rigid body. This is the rigid body that is connected to
+   * this rigid body's joint.
+   *
+   * @param[in] parent A pointer to this rigid body's parent rigid body.
+   */
+  void set_parent(RigidBody* parent);
+
+  /**
+   * Returns a const pointer to this rigid body's parent rigid body.
+   */
+  const RigidBody* get_parent() const;
+
+  /**
+   * Returns a mutable pointer to this rigid body's parent rigid body.
+   */
+  RigidBody* get_mutable_parent();
+
   bool hasParent() const;
 
   /**
@@ -87,7 +105,9 @@ class DRAKERBM_EXPORT RigidBody {
    * @return true if the supplied rigid body parameter other is the parent of
    * this rigid body.
    */
-  bool has_as_parent(const RigidBody& other) const { return parent == &other; }
+  bool has_as_parent(const RigidBody& other) const {
+    return parent_ == &other;
+  }
 
   void addVisualElement(const DrakeShapes::VisualElement& elements);
 
@@ -155,10 +175,6 @@ class DRAKERBM_EXPORT RigidBody {
   // note: it's very ugly, but parent, dofnum, and pitch also exist currently
   // (independently) at the RigidBodyTree level to represent the featherstone
   // structure.  this version is for the kinematics.
-
-  // TODO(amcastro-tri): Make it private and change to parent_.
-  /// The rigid body that's connected to this rigid body's joint.
-  RigidBody* parent;
 
   /// The index of this rigid body in the rigid body tree.
   int body_index;
@@ -228,4 +244,7 @@ class DRAKERBM_EXPORT RigidBody {
 
   // A unique ID for each model. It uses 0-index, starts from 0.
   int model_id_;
+
+  // The rigid body that's connected to this rigid body's joint.
+  RigidBody* parent_;
 };